+++ src/lib/commands.ts
@@ -2,8 +2,8 @@ import * as vscode from 'vscode'
 import { get_config, set_config, show_file } from './config'
 import { launch_renpy, launch_sdk } from './launch'
 import { prompt_configure_extensions } from './onboard'
-import { get_sdk_path, find_projects_in_workspaces } from './path'
-import { prompt_sdk_quick_pick } from './sdk'
+import { find_projects_in_workspaces } from './path'
+import { get_sdk_path, prompt_sdk_quick_pick } from './sdk'
 import { prompt_install_rpe, uninstall_rpes } from './rpe'
 import { get_executable } from './sh'
 import { WarpSocketService } from './socket'
@@ -205,7 +205,7 @@ export function get_commands(
 			const executable = await get_executable(sdk_path)
 			if (!executable) return
 
-			const projects = await find_projects_in_workspaces(context)
+			const projects = await find_projects_in_workspaces()
 			for (const project_root of projects) {
 				await prompt_install_rpe({
 					project: project_root,
